task-local random numbers; not deterministic, though
the hope was to achieve reproducible outcomes in multi-threaded runs 
with dynamic scheduling by attaching a random number generator to each 
task. The outcome was not achieved, as the order of iteration over a 
dict apparently depends on the order of adding to the dict, which is 
non-deterministic due to the dynamic scheduler.

fixed bug in adding stats; made stats Int everywhere